requirements.txt
python-telegram-bot
zxing-cpp
Pillow
python-dotenv
